finalized styling and cleaning up code
