+++ package.json
@@ -10,6 +10,7 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.24.0",
+    "firebase": "^9.6.3",
     "mini-css-extract-plugin": "2.4.7",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
