+++ src/App.js
@@ -1,11 +1,11 @@
 import { useEffect, useState } from 'react'
 import axios from 'axios';
-import DisplayResults from './DisplayResults.js';
-import InputForm from './InputForm.js';
-import Popup from './Popup.js';
-import SearchAgainButton from './SearchAgainButton'
-
-// import './App.css';
+import Header from './components/Header.js'
+import DisplayResults from './components/DisplayResults.js';
+import InputForm from './components/InputForm.js';
+import SearchAgainButton from './components/SearchAgainButton'
+import Footer from './components/Footer.js';
+import Popup from './components/Popup.js';
 
 function App() {
   const [nasaItem, setNasaItem] = useState([]);
@@ -15,14 +15,15 @@ function App() {
   const [searchAgain, setSearchAgain] = useState(false);
   
   const handleInput = (event) => {
-    setUserInput(event.target.value)
-  }
+    setUserInput(event.target.value);
+  };
   
   const handleLike = (id) => {
     const newNasaItem = [...nasaItem];
     const index = newNasaItem.findIndex(item => item.data[0].nasa_id === id);
-    newNasaItem[index].likes = !newNasaItem[index].likes
+    newNasaItem[index].likes = !newNasaItem[index].likes;
     setNasaItem(newNasaItem);
+    console.log('nasaItem', nasaItem);
   }
 
   const handleSubmit = (event) => {
@@ -30,6 +31,7 @@ function App() {
     setSearchTerm(userInput);
     setUserInput(''); 
   }
+  // End of Handle Functions
 
   useEffect (() => {
     if (searchTerm !== '') {
@@ -43,28 +45,25 @@ function App() {
       },
     }).then((response) => {
       const shorterResponse = (response.data.collection.items).slice(0, 9)
-            
       const createLikes = (nasaItem) => {
         const responseWithLikes = [...nasaItem];
         responseWithLikes.forEach(item => {
           item.likes = false
         });
-        
-        setNasaItem(responseWithLikes)
+      setNasaItem(responseWithLikes)
       }
-        shorterResponse.length === 0 ? setPopup(true) : createLikes(shorterResponse);
-    })
+      shorterResponse.length === 0 ? setPopup(true) : createLikes(shorterResponse);
+    }).catch(err => (console.log(err)));
     }
-  }, [searchTerm])
+  }, [searchTerm]);
+  // End of getting data from API and storing in the nasaItem
+  
 
   return (
     <div className='mainContainer'>
-      <header>
-        <div className='wrapper headerContainer'>
-          <h1 id='home'>A look at Space: <span>Images from NASA</span></h1>
-          <p>We choose to go to the moon in this decade and do the other things, not because they are easy, but because they are hard, because that goal will serve to organize and measure the best of our energies and skills, because that challenge is one that we are willing to accept, one we are unwilling to postpone, and one which we intend to win.</p>
-        </div>
-      </header>
+
+      <Header />
+
       <InputForm 
         handleSubmit={handleSubmit}
         handleInput={handleInput}
@@ -72,7 +71,7 @@ function App() {
         setSearchAgain={setSearchAgain}
       />
 
-    <section className='wrapper resultsContainer'>
+    <main className='wrapper resultsContainer'>
       {
         nasaItem.map((nasaObject) => {
           return(
@@ -86,18 +85,17 @@ function App() {
                 handleLike={handleLike}
                 />
               )
-        })
+          })
       }
+      {/* End of Displayed Data */}
 
       <SearchAgainButton 
         searchAgain={searchAgain}
         setSearchAgain={setSearchAgain}
         />
-    </section>
+    </main>
 
-    <footer>
-      <p>Created at <a href='htts://www.junocollege.com'>Juno College</a> by Ashley Thadickal</p>
-    </footer>
+      <Footer />
 
       <Popup
         trigger={popup}
