+++ src/components/Footer.js
@@ -1,0 +1,11 @@
+// Footer.js
+
+const Footer = () => {
+    return (
+        <footer>
+            <p>Created at <a href='htts://www.junocollege.com'>Juno College</a> by Ashley Thadickal</p>
+        </footer>
+    )
+}
+
+export default Footer
