+++ src/components/Header.js
@@ -1,0 +1,14 @@
+// Header.js
+
+const Header = () => {
+    return (
+        <header>
+            <div className='wrapper headerContainer'>
+                <h1 id='home'>A look at Space: <span>Images from NASA</span></h1>
+                <p>Start your space exploration here; no telescope or rockets required. Drop your favourite galatic word below.</p>
+            </div>
+        </header>
+    )
+}
+
+export default Header
