+++ src/components/InputForm.js
@@ -3,8 +3,8 @@
 const InputForm = (props) => {
     return(
         <form onSubmit={props.handleSubmit}>
-            <label htmlFor="search">Enter your search for a fantastic voyage:</label>
-            <input type="text" id="search" onChange={props.handleInput} value={props.userInput} />
+            <label htmlFor='search'>Search here for your fantastic voyage:</label>
+            <input type='text' id='search' onChange={props.handleInput} value={props.userInput} />
             <button onClick={() => props.setSearchAgain(true)}>Search</button>
         </form>
     )
