+++ src/components/SearchAgainButton.js
@@ -1,11 +1,9 @@
-
+// SearchAgainButton.js
 
 const SearchAgainButton = (props) => {
     return (props.searchAgain) ? (
         <button className="searchAgain" onClick={() => {props.setSearchAgain(false); }}><a href="#home">Click to Search Again</a></button>
-    ) : ('')
-        
-    
+    ) : ('')   
 }
 
 export default SearchAgainButton
