+++ src/index.css
@@ -2,12 +2,23 @@ html{line-height:1.15;-ms-text-size-adjust:100%;-webkit-text-size-adjust:100%}bo
 
 * { -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box; }
 
-.sr-only { position: absolute; width: 1px; height: 1px; margin: -1px; border: 0; padding: 0; white-space: nowrap; clip-path: inset(100%); clip: rect(0 0 0 0); overflow: hidden;}
+/* Typography and General Styling */
+
+@import url('https://fonts.googleapis.com/css2?family=Source+Code+Pro:wght@400;600&display=swap');
 
 html,
 body {
   height: 100%;
   position: relative;
+}
+
+html {
+  font-size: 125%;
+  font-family: 'Source Code Pro', monospace;
+  }
+
+body {
+  background-color: #000;
 }
 
 .mainContainer {
@@ -17,46 +28,10 @@ body {
   padding-bottom: 100px;
 }
 
-html {
-  font-size: 125%;
-  }
-
-body {
-  background-color: #000;
-}
-
 .wrapper {
   max-width: 1200px;
   width: 80%;
   margin: 0 auto;
-}
-
-.headerContainer {
-  padding-top: 150px;
-}
-
-h1 {
-  font-size: 2rem;
-  text-align: left;
-  color: white;
-}
-
-h1 span {
-  display: block;
-}
-
-h2 {
-  font-size: 1rem;
-}
-
-/* p {
-  padding: 0 20px;
-} */
-
-header p {
-  color: white;
-  width: 60%;
-  padding: 0;
 }
 
 button {
@@ -74,18 +49,66 @@ a {
   color: #000;
 }
 
+/* Header Styling */
+.headerContainer {
+  padding-top: 150px;
+}
+
 header {
   background-image: url(./19849.jpg);
-  background-size: cover;
+  background-size: 100%;
   background-repeat: no-repeat;
-  background-position: center;
+  background-position: bottom;
   height: 60vh;
 }
 
+h1 {
+  font-size: 2rem;
+  text-align: left;
+  color: white;
+}
+
+h1 span {
+  display: block;
+}
+
+header p {
+  color: white;
+  width: 55%;
+  line-height: 1.4;
+}
+
+/* Form Styling */
+form {
+  background-image: linear-gradient(-45deg,#fcda7e,#ff9733, #da4014);
+  background-size: 400%;
+  animation: gradient 15s ease infinite;
+  height: 150px;
+  width: 50%;
+  margin: -50px auto 10px auto;
+  border-radius: 5px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-evenly;
+  }
+  
+@keyframes gradient {
+  0% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+  100% {
+    background-position: 0% 50%;
+  }
+}
+
+/* Results Display Styling */
 .resultsContainer {
   display: flex;
-  flex-wrap: wrap;
- 
+  flex-wrap: wrap; 
 }
 
 .results {
@@ -95,10 +118,19 @@ header {
   padding: 8px;
 }
 
+h2 {
+  font-size: 1rem;
+}
+
 .results img {
   max-width: 100%;
 }
 
+.unliked {
+  color: red;
+}
+
+/* Search Again Button Styling */
 .searchAgain {
   background: #d1cde4;
   display: block;
@@ -107,37 +139,8 @@ header {
   height: 50px
 }
 
-
-form {
-  background-image: linear-gradient(-45deg,#fcda7e,#ff9733, #da4014);
-  background-size: 400%;
-  animation: gradient 15s ease infinite;
-  height: 150px;
-  width: 50%;
-  margin: -15px auto 10px auto;
-  border-radius: 5px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-evenly;
-  }
-  
-  @keyframes gradient {
-    0% {
-      background-position: 0% 50%;
-    }
-    50% {
-      background-position: 100% 50%;
-    }
-    100% {
-      background-position: 0% 50%;
-    }
-  }
-.unliked {
-  color: red;
-}
-
- .popup {
+/* Pupup Styling */
+.popup {
   position: fixed;
   top: 0;
   left: 0;
@@ -162,18 +165,20 @@ form {
   right: 16px;
 }
 
+/* Footer Styling */
 footer {
   color: white;
   position: absolute;
   bottom: 0%;
-  width: 100%;
-  left: 20px;
+  width: 95%;
+  left: 50px;
 }
 
 footer a {
   color: white;
 }
 
+/* Media Queries */
 @media (max-width: 940px) {
   .results {
     width: calc(50% - 20px);
@@ -196,7 +201,7 @@ footer a {
   }
 
   form {
-    margin-top: -60px;
+    margin-top: 10px;
     width: 70%;
   }
 
@@ -209,12 +214,12 @@ footer a {
   }
 }
 
-@media (max-width: 385px) {
+@media (max-width: 475px) {
   header {
     height: 40vh;
   }
   .headerContainer {
-    padding-top: 150px;
+    padding-top: 120px;
   }
   header p {
     display: none;
